iot-app/flask-server/requirements.txt
Flask==1.1.2
Flask_Cors==3.0.10
keras==2.11.0
numpy==1.22.4
opencv_python==4.7.0.68
Pillow==9.4.0
scikit_image==0.18.1
skimage==0.0
tensorflow==2.11.0
tensorflow_intel==2.11.0
gunicorn 